Update AlwaysInstallElevated Check
The first change (line 237) adds another link to the comment. The original source (FuzzySecurity) used the GreyHatHacker reference, and it talks about how both registry keys need to be set for the system to be vulnerable. 
The second change (line 248) is because the registry key should check the Local Machine (HKLM) and Current User (HKCU), and currently both of them check HKLM.
The third change (line 254) is to check for both keys to be set instead of just one key be set.